.agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Governed workflow CLI for the Inspected application"
authors = [
    {name = "Justin Cook", email = "jhcook@secnix.com"},
]
dependencies = [
    "typer>=0.21.1",
    "rich>=14.2.0",
    "pydantic>=2.12.5",
    "openai>=2.15.0",
    "google-genai>=1.57.0",
    "PyYAML>=6.0.2",
    "tiktoken>=0.12.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache-2.0"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
